+++ DotNetCore_React.Application/UserApp/Dtos/UserSimpleDto.cs
@@ -12,7 +12,7 @@ namespace DotNetCore_React.Application.UserApp.Dtos
 
         public string LastName { get; set; }
 
-        public byte Status { get; set; }
+        public int Status { get; set; }
 
     }
 }
